Do not import Qt on `__init__.py` so that `addon.py` does not need to load it either
