+++ client/ayon_ui_qt/components/check_box.py
@@ -6,7 +6,7 @@ from PySide6.QtCore import QRect, QSize
 from qtpy.QtGui import QPainter, QPaintEvent
 from qtpy.QtWidgets import QCheckBox, QSizePolicy, QStyle, QStyleOptionButton
 
-from .. import get_ayon_style
+from ..ayon_style import get_ayon_style
 from ..variants import QCheckBoxVariants
