+++ client/ayon_ui_qt/components/scroll_area.py
@@ -11,7 +11,7 @@ from qtpy.QtWidgets import (
     QStyleOptionSlider,
 )
 
-from .. import get_ayon_style
+from ..ayon_style import get_ayon_style
 
 
 class AYScrollBar(QScrollBar):
