+++ client/ayon_ui_qt/components/tag.py
@@ -7,7 +7,7 @@ try:
 except ImportError:
     from ..vendor.qtmaterialsymbols import get_icon
 
-from .. import get_ayon_style
+from ..ayon_style import get_ayon_style
 from .buttons import AYButton
 from .frame import AYFrame
 from .label import AYLabel
